refactor: yapıda genel düzenleme çalışmaları yapıldı.
