--- parsql-core/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "parsql-core"
-authors.workspace = true
-description.workspace = true
-edition.workspace = true
-categories.workspace = true
-keywords.workspace = true
-repository.workspace = true
-version.workspace = true
-license.workspace = true
-
-[features]
-default = []
-sqlite = ["dep:rusqlite"]
-postgres = ["dep:postgres"]
-tokio-postgres = ["dep:tokio-postgres"]
-deadpool-postgres = ["dep:tokio-postgres", "dep:deadpool-postgres"]
-
-[dependencies]
-rusqlite = { version = "0.33.0", optional = true }
-postgres = { version = "0.19.10", optional = true }
-tokio-postgres = { version = "0.7.13", optional = true }
-deadpool-postgres = { version = "0.14.1", optional = true }
-
-[lints]
-workspace = true
