+++ parsql-postgres/Cargo.toml
@@ -10,11 +10,9 @@ repository.workspace = true
 [dependencies]
 postgres = { version = "0.19.10" }
 
-[dependencies.parsql-core]
-workspace = true
-
 [dependencies.parsql-macros]
 workspace = true
+features = ["postgres"]
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
