+++ parsql-sqlite/Cargo.toml
@@ -8,13 +8,11 @@ authors.workspace = true
 repository.workspace = true
 
 [dependencies]
-rusqlite = { version = "0.33.0" }
-
-[dependencies.parsql-core]
-workspace = true
+rusqlite = { version = "0.34.0" }
 
 [dependencies.parsql-macros]
 workspace = true
+features = ["sqlite"]
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
