+++ parsql-sqlite/src/lib.rs
@@ -1,8 +1,5 @@
 pub mod crud_ops;
 
-// Re-export traits from parsql-core
-pub use parsql_core::{SqlQuery, SqlParams, UpdateParams, FromRow};
-
 // Re-export sqlite types that might be needed
 pub use rusqlite::{Connection, Error, Row};
 pub use rusqlite::types::ToSql;
@@ -11,3 +8,18 @@ pub use rusqlite::types::ToSql;
 pub use crud_ops::{insert, select, select_all, update, delete, get, get_all};
 
 pub use parsql_macros as macros;
+
+pub trait SqlQuery {
+    fn query() -> String;
+}
+pub trait SqlParams {
+    fn params(&self) -> Vec<&(dyn ToSql + Sync)>;
+}
+pub trait UpdateParams {
+    fn params(&self) -> Vec<&(dyn ToSql + Sync)>;
+}
+pub trait FromRow {
+    fn from_row(row: &Row) -> Result<Self, Error>
+    where
+        Self: Sized;
+}
