+++ parsql-tokio-postgres/Cargo.toml
@@ -17,11 +17,9 @@ postgres = { version = "0.19.10" }
 tokio-postgres = { version = "0.7.13" }
 deadpool-postgres = { version = "0.14.1", optional = true }
 
-[dependencies.parsql-core]
-workspace = true
-
 [dependencies.parsql-macros]
 workspace = true
+features = ["tokio-postgres"]
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports", "async_tokio"] }
