+++ parsql-tokio-postgres/src/lib.rs
@@ -1,8 +1,5 @@
 pub mod crud_ops;
 
-// Re-export traits from parsql-core
-pub use parsql_core::{SqlQuery, SqlParams, UpdateParams, FromRow};
-
 // Re-export tokio-postgres types that might be needed
 pub use tokio_postgres::{types::ToSql, Row, Error, Client};
 
@@ -18,6 +15,21 @@ pub use crud_ops::{
 };
 
 pub use parsql_macros as macros;
+
+pub trait SqlQuery {
+    fn query() -> String;
+}
+pub trait SqlParams {
+    fn params(&self) -> Vec<&(dyn ToSql + Sync)>;
+}
+pub trait UpdateParams {
+    fn params(&self) -> Vec<&(dyn ToSql + Sync)>;
+}
+pub trait FromRow {
+    fn from_row(row: &Row) -> Result<Self, Error>
+    where
+        Self: Sized;
+}
 
 #[cfg(feature = "deadpool-postgres")]
 pub mod transactional_ops;
